requirements.txt
asgiref==3.5.0
autopep8==1.6.0
Babel==2.9.1
black==22.3.0
CacheControl==0.12.11
cachetools==5.2.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
cloudinary==1.29.0
dj-database-url==0.5.0
Django==3.2.12
django-cors-headers==3.11.0
django-filter==21.1
django-on-heroku==1.1.2
django-prices==2.2.0
django-versatileimagefield==2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
enmerkar==0.7.1
firebase-admin==5.2.0
google-api-core==2.8.2
google-api-python-client==2.51.0
google-auth==2.8.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.1
google-cloud-firestore==2.5.3
google-cloud-storage==2.4.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.2
grpcio==1.46.3
grpcio-status==1.46.3
httplib2==0.20.4
idna==3.3
importlib-metadata==4.11.3
Markdown==3.3.6
msgpack==1.0.4
mypy-extensions==0.4.3
mysqlclient==2.1.0
pathspec==0.9.0
Pillow==9.0.1
platformdirs==2.5.2
prices==1.1.0
proto-plus==1.20.6
protobuf==3.20.1
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
PyJWT==2.3.0
PyMySQL==1.0.2
pyparsing==3.0.9
python-dotenv==0.20.0
python-magic==0.4.25
pytz==2021.3
requests==2.27.1
rsa==4.8
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
typing-extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.9
whitenoise==6.0.0
zipp==3.7.0
